pyproject.toml
[tool.poetry]
name = "jaundice-rate"
version = "0.1.0"
description = ""
authors = ["Sergei Klimov <s.klimov@bk.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
aiohttp = "^3.8.4"
beautifulsoup4 = "^4.12.2"
requests = "^2.31.0"
pymorphy2 = "^0.9.1"
pytest = "^7.3.1"
anyio = "^3.7.0"
aiologger = "^0.7.0"
aiofile = "^3.8.7"
pytest-asyncio = "^0.21.0"
pytest-trio = "^0.8.0"
pytest-aiohttp = "^1.0.4"
setuptools = "^67.8.0"
python-dotenv = "^1.0.0"
aiocache = {extras = ["redis"], version = "^0.12.1"}
redis = "^4.5.5"

[tool.poetry.dev-dependencies]
blue = "^0.9.1"

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -s -vvv"
env = [
    "REDIS_STORAGE_TIME = 1"
]
testpaths = [
    "tests",
    "integration",
]
asyncio_mode = "auto"
filterwarnings = [
    "ignore::UserWarning",
    "ignore::DeprecationWarning",
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
